Decompose InspectorPanel.tsx into per-tab panels (#38)
Split the 2,917-line frontend/src/components/editor/InspectorPanel.tsx
into a compositional shell plus per-tab panel components under
frontend/src/components/editor/panels/.

- InspectorPanel.tsx is now 112 lines, purely compositional: it owns
  tab navigation, the export action, and composition of the five tab
  panels.
- One component per existing tab: FilesPanel, SourcesPanel,
  MemoryPanel, UsagePanel, TurnsPanel. Each panel reads only the
  context slice it needs from useApp() and owns its own local state
  (preview errors, usage fetch, memory drafts, source metadata cache).
- Shared helpers split into focused modules: types.ts, shared-utils.ts,
  primitives.tsx, memory-document.ts, artifacts-data.tsx,
  sources-data.tsx. No behavior changes.
- Per-panel Vitest smoke tests (FilesPanel, SourcesPanel, MemoryPanel,
  UsagePanel, TurnsPanel) plus a shared test/panel-fixtures.ts that
  builds a mock AppContextValue.
- Kept the single React Context (per CLAUDE.md's "no extra state
  library" rule): each panel owns state locally rather than receiving
  its own store slice.

Issue proposed names (Evidence/Artifacts/Memory/Skills/Usage) were
mapped to the actual tab structure (Files/Sources/Memory/Usage/Turns);
there is no Skills tab in the inspector today. The proposed Playwright
visual-regression baselines are out of scope without a snapshot
harness in frontend/e2e; this change keeps the existing Playwright
suite untouched.

Verified with: npm run typecheck, npm test (115 passed), npm run lint.
