+++ frontend/src/components/editor/panels/artifacts-data.tsx
@@ -1,0 +1,244 @@
+"use client";
+
+import { FileText } from "lucide-react";
+import {
+  getPreviewableFileLabel,
+  inferPreviewableFileKind,
+} from "@/lib/file-preview";
+import { getMessageToolCalls } from "@/lib/message-blocks";
+import type { Message } from "@/lib/types";
+import { cn } from "@/lib/utils";
+import {
+  humanizeLabel,
+  humanizeToken,
+  shortenPath,
+} from "./shared-utils";
+import type { GeneratedArtifactItem, GeneratedArtifactKind } from "./types";
+
+export function shouldShowGeneratedArtifact(item: {
+  path: string;
+  artifactType: string | null;
+}): boolean {
+  if (!item.path) {
+    return false;
+  }
+
+  if (item.artifactType?.endsWith("_run")) {
+    return false;
+  }
+
+  return true;
+}
+
+export function collectArtifacts(messages: Message[]): GeneratedArtifactItem[] {
+  const items = new Map<string, GeneratedArtifactItem>();
+  let order = 0;
+
+  const upsertArtifact = ({
+    path,
+    artifactType,
+    sourceTool,
+  }: {
+    path: string;
+    artifactType?: string | null;
+    sourceTool?: string | null;
+  }) => {
+    const existing = items.get(path);
+    const nextItem: GeneratedArtifactItem = {
+      path,
+      label: path.split("/").pop() ?? path,
+      artifactType: artifactType ?? existing?.artifactType ?? null,
+      sourceTool: sourceTool ?? existing?.sourceTool ?? null,
+      lastSeenOrder: order,
+    };
+    order += 1;
+
+    if (!shouldShowGeneratedArtifact(nextItem)) {
+      return;
+    }
+
+    items.set(path, nextItem);
+  };
+
+  messages.forEach((message) => {
+    getMessageToolCalls(message).forEach((call) => {
+      call.result?.artifact_refs?.forEach((artifact) => {
+        if (!artifact.path) {
+          return;
+        }
+
+        upsertArtifact({
+          path: artifact.path,
+          artifactType: artifact.artifact_type ?? null,
+          sourceTool: call.tool,
+        });
+      });
+    });
+  });
+
+  return Array.from(items.values())
+    .sort((left, right) => right.lastSeenOrder - left.lastSeenOrder)
+    .slice(0, 12);
+}
+
+function inferGeneratedArtifactKind(item: GeneratedArtifactItem): GeneratedArtifactKind {
+  return inferPreviewableFileKind({
+    path: item.path,
+    artifactType: item.artifactType,
+    label: item.label,
+  });
+}
+
+function getGeneratedArtifactCue(item: GeneratedArtifactItem) {
+  const kind = inferGeneratedArtifactKind(item);
+  return {
+    kind,
+    label: getPreviewableFileLabel({
+      path: item.path,
+      artifactType: item.artifactType,
+      label: item.label,
+    }),
+  };
+}
+
+function getGeneratedArtifactTone(kind: GeneratedArtifactKind) {
+  if (kind === "table") {
+    return {
+      badge: "border-amber-200 bg-amber-50 text-amber-700",
+      icon: "bg-amber-50 text-amber-700",
+      active:
+        "border-amber-200 bg-[linear-gradient(180deg,rgba(255,251,235,0.95),rgba(254,243,199,0.72))]",
+      idle: "border-[rgba(245,158,11,0.14)] bg-[rgba(255,251,235,0.68)]",
+    };
+  }
+
+  if (kind === "plot") {
+    return {
+      badge: "border-rose-200 bg-rose-50 text-rose-700",
+      icon: "bg-rose-50 text-rose-700",
+      active:
+        "border-rose-200 bg-[linear-gradient(180deg,rgba(255,241,242,0.96),rgba(255,228,230,0.74))]",
+      idle: "border-[rgba(244,63,94,0.12)] bg-[rgba(255,244,246,0.7)]",
+    };
+  }
+
+  if (kind === "report") {
+    return {
+      badge: "border-emerald-200 bg-emerald-50 text-emerald-700",
+      icon: "bg-emerald-50 text-emerald-700",
+      active:
+        "border-emerald-200 bg-[linear-gradient(180deg,rgba(236,253,245,0.96),rgba(209,250,229,0.72))]",
+      idle: "border-[rgba(16,185,129,0.12)] bg-[rgba(240,253,244,0.72)]",
+    };
+  }
+
+  if (kind === "structured") {
+    return {
+      badge: "border-sky-200 bg-sky-50 text-sky-700",
+      icon: "bg-sky-50 text-sky-700",
+      active:
+        "border-sky-200 bg-[linear-gradient(180deg,rgba(240,249,255,0.96),rgba(224,242,254,0.74))]",
+      idle: "border-[rgba(14,165,233,0.12)] bg-[rgba(240,249,255,0.7)]",
+    };
+  }
+
+  if (kind === "archive") {
+    return {
+      badge: "border-slate-200 bg-slate-100 text-slate-600",
+      icon: "bg-slate-100 text-slate-600",
+      active:
+        "border-slate-300 bg-[linear-gradient(180deg,rgba(248,250,252,0.96),rgba(241,245,249,0.82))]",
+      idle: "border-[rgba(148,163,184,0.14)] bg-[rgba(248,250,252,0.78)]",
+    };
+  }
+
+  return {
+    badge: "border-stone-200 bg-stone-50 text-stone-700",
+    icon: "bg-stone-50 text-stone-700",
+    active:
+      "border-stone-200 bg-[linear-gradient(180deg,rgba(250,250,249,0.96),rgba(245,245,244,0.82))]",
+    idle: "border-[rgba(168,162,158,0.14)] bg-[rgba(250,250,249,0.76)]",
+  };
+}
+
+function getGeneratedArtifactDetail(item: GeneratedArtifactItem): string {
+  const values = [humanizeToken(item.artifactType), humanizeLabel(item.sourceTool)].filter(
+    (value): value is string => Boolean(value)
+  );
+
+  const uniqueValues = values.filter(
+    (value, index) => values.indexOf(value) === index
+  );
+
+  return uniqueValues[0] ?? "Generated artifact";
+}
+
+function getGeneratedArtifactScopeLabel(item: GeneratedArtifactItem): string | null {
+  return humanizeLabel(item.sourceTool);
+}
+
+export function GeneratedFileRow({
+  item,
+  active,
+  onClick,
+}: {
+  item: GeneratedArtifactItem;
+  active: boolean;
+  onClick: () => void;
+}) {
+  const cue = getGeneratedArtifactCue(item);
+  const tone = getGeneratedArtifactTone(cue.kind);
+  const detail = getGeneratedArtifactDetail(item);
+  const scopeLabel = getGeneratedArtifactScopeLabel(item);
+
+  return (
+    <button
+      type="button"
+      onClick={onClick}
+      title={item.path}
+      className={cn(
+        "flex w-full items-start gap-2 rounded-[12px] border px-2.5 py-2 text-left transition-colors",
+        active
+          ? tone.active
+          : `${tone.idle} hover:border-[rgba(211,219,210,0.9)] hover:bg-white`
+      )}
+    >
+      <span
+        className={cn(
+          "mt-0.5 flex h-7 w-7 flex-shrink-0 items-center justify-center rounded-[9px]",
+          tone.icon
+        )}
+      >
+        <FileText size={12} />
+      </span>
+      <span className="min-w-0 flex-1">
+        <span className="flex items-start justify-between gap-2">
+          <span className="min-w-0 truncate text-[12px] font-semibold text-slate-700">
+            {item.label}
+          </span>
+          <span
+            className={cn(
+              "inline-flex shrink-0 items-center rounded-full border px-1.5 py-0.5 text-[9px] font-semibold uppercase tracking-[0.08em]",
+              tone.badge
+            )}
+          >
+            {cue.label}
+          </span>
+        </span>
+        <span className="mt-1 flex min-w-0 items-center gap-1.5">
+          <span className="truncate text-[10px] font-medium text-slate-500">
+            {detail}
+          </span>
+          {scopeLabel ? (
+            <span className="shrink-0 rounded-full bg-white/80 px-1.5 py-0.5 text-[9px] font-medium uppercase tracking-[0.08em] text-slate-400">
+              {scopeLabel}
+            </span>
+          ) : null}
+        </span>
+        <span className="mt-1 block truncate font-mono text-[9px] text-slate-400">
+          {shortenPath(item.path, 4)}
+        </span>
+      </span>
+    </button>
+  );
+}
